refactoring protocol step 1

+++ app/model/PitchRollHeading.scala
@@ -3,4 +3,4 @@ package model
 /**
  * Created by becker on 01.08.14.
  */
-case class PitchRollHeading(pitch: Float, roll: Float, trueHeading: Float, magHeading: Float)
+case class PitchRollHeading(pitch: BigDecimal, roll: BigDecimal, trueHeading: BigDecimal, magHeading: BigDecimal)
